Create one build file at the top of the project.
Just more in line with the way Kobalt manages projects: put them all in one build file. Easier to share values, add dependencies between projects (e.g. consumer should depend on colaTestsPlugin), etc...
